add activity logging and enhance expense management; implement activity routes and update models

+++ src/controllers/expenseController.js
@@ -1,6 +1,7 @@
 import {
     Expense,
-    ExpenseParticipant
+    ExpenseParticipant,
+    ActivityLog
 } from "../models/index.js";
 
 import { calculateBalances } from "../services/balanceService.js";
@@ -12,29 +13,49 @@ export const addExpense = async (req, res) => {
         const {
             description,
             amount,
+            currency,
             paidBy,
-            participants
+            participants,
+            expenseDate
         } = req.body;
+
+        //edge case
+        if(!participants || participants.length === 0) {
+            return res.status(400).json({
+                message: "At least one participant is required"
+            });
+        }
 
         const expense = await Expense.create({
             description,
             amount,
-            paidBy
+            currency,
+            paidBy,
+            createdBy : req.user.userId,
+            expenseDate
         });
 
-        const share = amount / participants.length;
+        // const share = amount / participants.length; // gives x.3333 sometimes
+
+        const share = Number((amount / participants.length).toFixed(2)); // rounds to 2 decimal places
 
         for (const userId of participants) {
             console.log("creating participant", userId);
             await ExpenseParticipant.create({
                 expenseId: expense.expenseId,
                 userId,
-                share
+                amountOwed : share
             });
         }
 
+        await ActivityLog.create({
+            expenseId: expense.expenseId,
+            userId: req.user.userId,
+            action: "Expense Created",
+        });
+
         res.status(201).json({
-            message: "Expense added",
+            message: "Expense added successfully",
             expense
         });
 
@@ -49,7 +70,23 @@ export const addExpense = async (req, res) => {
 // GET ALL EXPENSES
 export const getAllExpenses = async (req, res) => {
     try {
-        const expenses = await Expense.findAll();
+        const expenses = await Expense.findAll({
+            include : [
+                {
+                    model: ExpenseParticipant,
+                    where: {
+                        userId: req.user.userId
+                    },
+                }
+            ]
+        });
+
+        //equivalent sql query: Get all expenses with the logged in user as participant
+        // SELECT * FROM expenses e 
+        // JOIN expense_participants ep 
+        // ON e.expenseId = ep.expenseId 
+        // WHERE ep.userId = req.user.userId;
+
         res.json(expenses);
     } catch (error) {
         res.status(500).json({
@@ -59,11 +96,20 @@ export const getAllExpenses = async (req, res) => {
 };
 
 
-// GET SINGLE EXPENSE
+// GET  EXPENSE BY ID
 export const getExpenseById = async (req, res) => {
     try {
-        const { id } = req.params;
-        const expense = await Expense.findByPk(id);
+        // const { id } = req.params;
+        const expense = await Expense.findByPk(
+            req.params.id,
+            {
+                include: [
+                    {
+                        model: ExpenseParticipant
+                    }
+                ]
+            }
+        );
 
         if (!expense) {
             return res.status(404).json({
@@ -110,7 +156,7 @@ export const getExpenseById = async (req, res) => {
 export const updateExpense = async (req,res) => {
     try {
         const { id } = req.params;
-        const {description, amount, paidBy, participants} = req.body;
+        const {description, amount, currency, paidBy, participants, expenseDate} = req.body;
 
         const expense = await Expense.findByPk(id);
 
@@ -120,11 +166,20 @@ export const updateExpense = async (req,res) => {
             });
         }
 
+        // new edge case: only the user who created the expense can update it
+        if(expense.createdBy !== req.user.userId){
+            return res.status(403).json({
+                message: "You are not authorized to update this expense"
+            });
+        }
+
         // updating the expense table
         await expense.update({
             description,
             amount,
-            paidBy
+            currency,
+            paidBy,
+            expenseDate
         });
 
         // delete old participants
@@ -135,16 +190,28 @@ export const updateExpense = async (req,res) => {
         });
 
         // add new participants
-        const share = amount / participants.length;
+        // const share = amount / participants.length;
+        const share = Number((amount / participants.length).toFixed(2));
         for(const userId of participants){
             await ExpenseParticipant.create({
                 expenseId: id,
                 userId,
-                share
+                amountOwed : share
             });
         }
+        
+        // add in the activity log that the expense was updated
+        await ActivityLog.create({
+            expenseId: id,
+            userId: req.user.userId,
+            action: "Expense Updated",
+        });
 
-        res.json(expense);
+        res.json({
+            message: "Expense updated successfully",
+             expense
+        });
+
     }   catch (error) {
         res.status(500).json({
             error: error.message
@@ -165,10 +232,23 @@ export const deleteExpense = async (req, res) => {
             });
         }
 
+        if(expense.createdBy !== req.user.userId){
+            return res.status(403).json({
+                message: "You are not authorized to update this expense"
+            });
+        }
+
+        // works like an audit log to keep track of all the actions performed by users on expenses
+        await ActivityLog.create({
+            expenseId: expense.expenseId,
+            userId: req.user.userId,
+            action: "Expense Deleted",
+        });
+
         // delete participants first
         await ExpenseParticipant.destroy({
             where: {
-                expenseId: id
+                expenseId: expense.expenseId
             }
         });
